refactoring and modify package-lock
