+++ bin/page-loader.js
@@ -6,8 +6,8 @@ import pageLoader from '../src/page-loader.js';
 program
   .version('0.0.1')
   .description('Simple CLI tool for download web page.')
-  .arguments('<pageUrl>')
+  .arguments('<url>')
   .option('-o, --output [path]', 'output path for download files', process.cwd())
-  .action((pageUrl) => pageLoader(new URL(pageUrl), program.output));
+  .action((url) => pageLoader(url, program.output));
 
 program.parse(process.argv);
