+++ src/page-loader.js
@@ -1,13 +1,42 @@
 import { promises as fsp } from 'fs';
+import cheerio from 'cheerio';
 import {
   convertLinkToFilename,
   convertLinkToDirname,
   buildPath,
-  processingResources,
+  buildResourceFilepath,
   loadData,
 } from './util.js';
 
-const pageLoader = (pageUrl, outputDir = process.cwd()) => {
+const tagAttributeMap = {
+  img: 'src',
+  link: 'href',
+  script: 'src',
+};
+
+const processingResources = (origin, htmlData, resourceDir) => {
+  const $ = cheerio.load(htmlData);
+  const tagWithResources = Object.keys(tagAttributeMap)
+    .flatMap((tagName) => $(tagName).toArray()
+      .map((tag) => {
+        const url = new URL($(tag).attr(tagAttributeMap[tagName]), origin);
+        return { tag, url };
+      })
+      .filter(({ url }) => url.origin === origin)
+      .map(({ tag, url }) => {
+        const filepath = buildResourceFilepath(url.hostname, resourceDir, url.pathname);
+        return { tag, url, filepath };
+      }));
+  tagWithResources.forEach(({ tag, filepath }) => {
+    const attrName = tagAttributeMap[tag.name];
+    $(tag).attr(attrName, filepath);
+  });
+  const resources = tagWithResources.map(({ url, filepath }) => ({ url, filepath }));
+  return { page: $.html(), resources };
+};
+
+const pageLoader = (pageLink, outputDir = process.cwd()) => {
+  const pageUrl = new URL(pageLink);
   const link = `${pageUrl.hostname}${pageUrl.pathname}`;
   const pageFilename = convertLinkToFilename(link);
   const pageFilepath = buildPath(outputDir, pageFilename);
